alterações no arquivo sql
