+++ src/models/client.js
@@ -1,9 +1,13 @@
 class Client {
-    constructor({clientCod, clientName, clientCPF, clientStatus, clientCreatedDate}){
+    constructor({clientEmail, clientPassword,  clientSalt, clientCod, clientName, clientCPF, clientStatus, clientCreatedDate}){
         if(!clientStatus){
             clientStatus = "Active"
         }
-        this.clientCod = clientCod,
+        this.id = id
+        this.clientEmail = clientEmail,
+        this.clientPassword = clientPassword,
+        this.clientSalt = clientSalt,
+        this.clientCod = clientCod, 
         this.clientName = clientName,
         this.clientCPF = clientCPF,
         this.clientStatus = clientStatus,
@@ -12,4 +16,5 @@ class Client {
 }
 
 module.exports = Client
+
