--- src/models/user.js
@@ -1,12 +1,0 @@
-
-class User {
-    constructor({id, username, password,salt}){
-       
-        this.id = id
-        this.username = username
-        this.password = password
-        this.salt = salt
-    }
-}
-
-module.exports = User;
