+++ src/repository/client.repository.js
@@ -1,16 +1,27 @@
 const database = require('../helpers/database.util.js')
-const Client = require('../models/client')
+const User = require('../models/user')
 
+const getUser = async (user) => {
+    return new Promise(async(resolve, reject) => {
+        try {
+            const sqlstatement = `SELECT * FROM user WHERE userName = "${user.username}" `
+            
+            const result = await database.query(sqlstatement)
+            resolve(result[0])
+
+        }catch(err) {
+            console.error(err)
+            reject(err)
+        }
+    })
+}
+/* 
 const newClient = async (client) => {
     const clientExists = await database.registerExists("client", "clientCPF", client.clientCPF)
     if(clientExists){
         return "Cliente já cadastrado"
     }
-/*clientCod,
-clientName,
-clientCPF,
-clientStatus,
-clientCreatedDate, */
+
     return new Promise(async(resolve, reject) => {
         try {
             const {clientName, clientCPF} = client
@@ -24,6 +35,11 @@ clientCreatedDate, */
             reject(err)
         }
     })
+} */
+
+
+
+
+module.exports = { 
+    getUser
 }
-
-
